Backend/requirements.txt
alembic==1.7.4
autopep8==1.5.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
inflect==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.0
psycopg2==2.9.1
pycodestyle==2.8.0
pydantic==1.8.2
pyparsing==3.0.1
python-dotenv==0.19.1
SQLAlchemy==1.4.26
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
flask-cors
